requirements.txt
requests
arrow==0.12.1
Django==2.0.7
djangorestframework==3.8.2
python-dateutil==2.7.3
pytz==2018.5
six==1.11.0
